+++ lib/active_record_extended/active_record.rb
@@ -18,12 +18,8 @@ require "active_record_extended/query_methods/inet"
 require "active_record_extended/query_methods/json"
 require "active_record_extended/query_methods/select"
 
-if ActiveRecord::VERSION::MAJOR == 5 && ActiveRecord::VERSION::MINOR <= 1
-  if ActiveRecord::VERSION::MINOR.zero?
-    require "active_record_extended/patch/5_0/regex_match"
-    require "active_record_extended/patch/5_0/predicate_builder_decorator"
-  end
+if ActiveRecord::VERSION::MAJOR == 5 && ActiveRecord::VERSION::MINOR == 1
   require "active_record_extended/patch/5_1/where_clause"
-elsif ActiveRecord::VERSION::MAJOR >= 5
-  require "active_record_extended/patch/5_2/where_clause"
+else
+  require "active_record_extended/patch/5_2/where_clause" # Works with Rail 6.0.x too
 end