+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Deprecations:
 - Dropping support for Ruby v2.3.x
+- Dropping support for Rails 5.0.x
 - Dropping support for `cast_as_array` option for json queries.
   - Same operation can be performed with `cast_with: :array`